requirements.txt
streamlit==1.29.0
sqlalchemy==2.0.23
pandas==2.1.4
plotly==5.18.0
openpyxl==3.1.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
